+++ lib/one_request.dart
@@ -5,4 +5,4 @@ export 'src/dio_request.dart';
 export 'src/model/error.dart';
 export 'src/resourses/types.dart';
 export 'src/resourses/utils.dart';
-//just call oneRequest and ready to go.
+//just call OneRequest and ready to go.